✅ 2R83M2 code: done
Task: 202604210859-2R83M2
Primary: code
Agent: CODER
Author: CODER
Status: DONE
Comment: verified: split Redmine backend runtime into context, state, operations, and methods modules | details: typecheck, targeted lint/format, backend-critical tests, and framework bootstrap passed.
